==> 01-data-pipelane.ipynb <==
# ML Process
1. Problem Definition
2. Data Preparation
3. EDA
4. Preprocessing
5. Training Model
6. Evaluation
7. API Services
8. PyTest
9. Deployment

### 1. Problem Definition

- Melakukan prediksi Global Sales berdasarkan informasi yang dimiliki, seperti Sales di masing - masing region nya.
- Harapannya kita bisa melihat most potential game based on Sales
- Pada case ini kita akan coba melakukan prediksi terhadap `Global_Sales`

Goals dari project ini untuk lebih kenal atau membiasakan dengan ML Process

Source data: https://www.kaggle.com/datasets/gregorut/videogamesales

import pandas as pd
import joblib
from src.utils.helper import load_params, read_data, check_data

%load_ext autoreload
%autoreload 2

params_dir = "config/params.yaml"

params = load_params(params_dir)

params

### 2. Data Collection 

data = read_data(params["dataset_dir"], "vgsales.csv")

data

data.info()

Drop columns `Rank`, `EU_Sales`, `JP_Sales`, `Other_Sales`, dan `Global_Sales`

data.drop(["Rank"], axis = 1, inplace = True)

data.head()

# save data into .pkl file
joblib.dump(data, "data/processed/data.pkl")

### 3. Data Definition

Name: Nama game
    [object]
    
Platform: Platform untuk games dirilis
    [object]
    [PC, PS4, PS2, XOne, WiiU, dll]

Year: Tahun game dirilis
    [integer]
    [1980 - 2020]

Genre: Genre dari game
    [object]
    [Sports, Platform, Racing, Puzzle, dll]

Publisher: Publisher yang publish game
    [object]
    [Mojang, Konami, EA, dll]

NA_Sales: Sales pada North America (in millions)
    [float]
    
EU_Sales: Sales pada Europe (in millions)
    [float]
    
JP_Sales: Sales pada Jepang (in millions)
    [float]
    
Other_Sales: Sales pada region lain (in millions)
    [float]

###  4. Data Validation

#### Check Data Types

# cek tipe data
data.dtypes

Agak aneh untuk column `Year` karena dalam bentuk float. Bisa kita ubah dalam bentuk int

#### Range 

data.describe()

#### Check data shape

data.shape

#### Check missing values 

data.isna().sum()

#### Handling Column "Year"  

# convert float into int

data["Year"] = data["Year"].astype("int").copy()

Ternyata ada missing values, sehingga kita tidak bisa melakukan casting

data["Year"].isnull().sum()

Kita bisa handle secara temporary dengan isi missing values tersebut dengan value `-1`

data["Year"].fillna(-1, inplace=True)

data["Year"].isnull().sum()

# Kita coba casting ke integer lagi
data["Year"] = data["Year"].astype("int").copy()

data.info()

#### Handling Column "Publisher"

data["Publisher"].isnull().sum()

Ada dua cara untuk treat data ini:
1. Kita drop aja missing values ini
2. Kita isi missing values ini dengan "UNKNOWN"

Tapi kita bisa treat di bagian preprocessing, karena disini kita cuma nge cek data aja

data[data["Publisher"].isnull() == True]

### 4. Data Defense 

check_data(data, params)

### 5. Data Splitting 

# Split into X and y
X = data[params["predictors"]].copy()
y = data[params["label"]].copy()

X.head()

y.head()

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, y,
                                                    test_size = 0.2,
                                                    random_state = 42)

X_valid, X_test, y_valid, y_test = train_test_split(X_test,
                                                    y_test,
                                                    test_size = 0.5,
                                                    random_state = 42)

joblib.dump(X_train, "data/processed/X_train.pkl")
joblib.dump(y_train, "data/processed/y_train.pkl")
joblib.dump(X_valid, "data/processed/X_valid.pkl")
joblib.dump(y_valid, "data/processed/y_valid.pkl")
joblib.dump(X_test, "data/processed/X_test.pkl")
joblib.dump(y_test, "data/processed/y_test.pkl")
