==> 03-preprocessing-feature-engineering.ipynb <==
import pandas as pd
import numpy as np
import joblib
from src.utils.helper import load_params, nan_detector

### ToDo List

1. Missing value handling
    - year : median / mode dari tahun
    - publisher : isi dengan value "UNKNOWN"
    
2. Treat outlier
    - remove outlier
    - IQR?
    - transform menggunakan log

### Import Params 

params_dir = "config/params.yaml"

params = load_params(params_dir)

params

### Load Dataset

X_train = joblib.load("data/processed/X_train.pkl")
y_train = joblib.load("data/processed/y_train.pkl")

X_valid = joblib.load("data/processed/X_valid.pkl")
y_valid = joblib.load("data/processed/y_valid.pkl")

X_test = joblib.load("data/processed/X_test.pkl")
y_test = joblib.load("data/processed/y_test.pkl")

data = pd.concat([X_train, y_train], axis=1)

valid_set = pd.concat([X_valid, y_valid], axis=1)

test_set = pd.concat([X_test, y_test], axis=1)

### Handling Missing Values 

#### Convert -1 value to NaN 

data.columns.tolist()

def coba(set_of_data):
    set_of_data = set_of_data.copy()
    
    columns = set_of_data.columns.tolist()
    
    for col in columns:
        if set_of_data[col].dtypes == "int":
            

for col in data.columns.tolist():
    if data[col].dtypes == "int":
        print("wikwik")
        
    else:
        print("awikwok")

##### Train

data.describe()

data.isna().sum()

data = nan_detector(data)

data.isna().sum()

data.info()

##### Validation

valid_set.describe()

valid_set.isnull().sum()

valid_set = nan_detector(valid_set)

valid_set.isna().sum()

##### Test 

test_set.describe()

test_set.isna().sum()

test_set = nan_detector(test_set)

test_set.isna().sum()

#### Handling Year 

##### Train 

impute_year = 2007

data["Year"].fillna(impute_year, inplace=True)

data.isna().sum()

##### Validation 

valid_set["Year"].fillna(impute_year, inplace=True)

valid_set.isna().sum()

##### Test 

test_set["Year"].fillna(impute_year, inplace=True)

test_set.isna().sum()

#### Handling Publisher column

##### Train 

data["Publisher"].fillna("UNKNOWN", inplace=True)

data.isna().sum()

##### Validation 

valid_set["Publisher"].fillna("UNKNOWN", inplace=True)

valid_set.isna().sum()

##### Test 

test_set["Publisher"].fillna("UNKNOWN", inplace=True)

test_set.isna().sum()

### Encoding

Sementara di skip dulu

from sklearn.preprocessing import OneHotEncoder

data["Platform"].nunique()

data["Genre"].nunique()

data["Publisher"].nunique()

data.drop(["Publisher"], axis=1, inplace = True)

ohe_genre = OneHotEncoder(sparse=False)

data["Genre"]

ohe_genre.fit(data["Genre"])

### Dump data 

cols_to_pick = ["Year", "NA_Sales", "EU_Sales", "JP_Sales", "Other_Sales", "Global_Sales"]

new_data = data[cols_to_pick].copy()

new_data

new_valid = valid_set[cols_to_pick].copy()

new_valid

new_test = test_set[cols_to_pick].copy()

joblib.dump(new_data.drop(columns = "Global_Sales"), "data/processed/X_train_new.pkl")
joblib.dump(new_valid.drop(columns = "Global_Sales"), "data/processed/X_valid_new.pkl")
joblib.dump(new_test.drop(columns = "Global_Sales"), "data/processed/X_test_new.pkl")
