==> 04-modelling.ipynb <==
from sklearn.linear_model import LinearRegression
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.neighbors import KNeighborsRegressor

from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
from sklearn.model_selection import RandomizedSearchCV

from datetime import datetime
import joblib
import json
from tqdm import tqdm
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import copy 
import hashlib

### Load Dataset 

X_train = joblib.load("data/processed/X_train_new.pkl")
y_train = joblib.load("data/processed/y_train.pkl")

X_valid = joblib.load("data/processed/X_valid_new.pkl")
y_valid = joblib.load("data/processed/y_valid.pkl")

X_test = joblib.load("data/processed/X_test_new.pkl")
y_test = joblib.load("data/processed/y_test.pkl")

### Create Log Template 

def time_stamp():
    return datetime.now()

def create_log_template():
    logger = {
        "model_name" : [],
        "model_uid" : [],
        "training_time" : [],
        "training_date" : [],
#         "performance" : [],
        "r2_score" : [],
        "mse_score": [],
        "mae_score": [],
        "data_configurations" : [],
    }

    return logger

def training_log_updater(current_log, log_path):
    current_log = current_log.copy()

    try:
        with open(log_path, "r") as file:
            last_log = json.load(file)
        file.close()
    except FileNotFoundError as ffe:
        with open(log_path, "w") as file:
            file.write("[]")
        file.close()
        with open(log_path, "r") as file:
            last_log = json.load(file)
        file.close()
    
    last_log.append(current_log)

    with open(log_path, "w") as file:
        json.dump(last_log, file)
        file.close()

    return last_log

### Training and Evaluation

#### Create Model Object 

lr_vanilla = LinearRegression()
dct_vanilla = DecisionTreeRegressor()
rfr_vanilla = RandomForestRegressor()
knn_vanilla = KNeighborsRegressor()

list_of_model = {
    "vanilla" : [
    { "model_name": lr_vanilla.__class__.__name__, "model_object": lr_vanilla, "model_uid": ""},
    { "model_name": dct_vanilla.__class__.__name__, "model_object": dct_vanilla, "model_uid": ""},
    { "model_name": rfr_vanilla.__class__.__name__, "model_object": rfr_vanilla, "model_uid": ""},
    { "model_name": knn_vanilla.__class__.__name__, "model_object": knn_vanilla, "model_uid": ""},
    ]
}

#### Baseline Model 

baseline_model = np.mean(y_train)
print(baseline_model)

baseline_mse = mean_squared_error(y_train, 
                                  np.ones(len(y_train)) * baseline_model)
print(baseline_mse)

#### Training Model 

def train_eval_model(list_of_model, prefix_model_name, X_train, y_train, data_configuration_name, X_valid, y_valid, log_path):

    list_of_model = copy.deepcopy(list_of_model)
    logger = create_log_template()

    for model in tqdm(list_of_model):    
        model_name = prefix_model_name + "-" + model["model_name"]

        start_time = time_stamp()
        model["model_object"].fit(X_train, y_train)
        finished_time = time_stamp()

        elapsed_time = finished_time - start_time
        elapsed_time = elapsed_time.total_seconds()

        y_pred = model["model_object"].predict(X_valid)
        performance_r2 = r2_score(y_valid, y_pred)
        performance_mse = mean_squared_error(y_valid, y_pred)
        performance_mae = mean_absolute_error(y_valid, y_pred)

        plain_id = str(start_time) + str(finished_time)
        chiper_id = hashlib.md5(plain_id.encode()).hexdigest()

        model["model_uid"] = chiper_id

        logger["model_name"].append(model_name)
        logger["model_uid"].append(chiper_id)
        logger["training_time"].append(elapsed_time)
        logger["training_date"].append(str(start_time))
#         logger["performance"].append(performance)
        logger["r2_score"].append(performance_r2)
        logger["mse_score"].append(performance_mse)
        logger["mae_score"].append(performance_mae)
        logger["data_configurations"].append(data_configuration_name)

    training_log = training_log_updater(logger, log_path)

    return training_log, list_of_model

training_log, list_of_model_vanilla = train_eval_model(
    list_of_model["vanilla"],
    "vanilla_model",
    X_train,
    y_train,
    "vanilla",
    X_valid,
    y_valid,
    "log/training_log.json"
)

list_of_model["vanilla"] = copy.deepcopy(list_of_model_vanilla)

### Best performance 

def training_log_to_df(training_log):
    training_res = pd.DataFrame()

    for log in tqdm(training_log):
        training_res = pd.concat([training_res, pd.DataFrame(log)])
    
    training_res.sort_values(["r2_score", "training_time"], ascending = [False, True], inplace = True)
    training_res.reset_index(inplace = True, drop = True)
    
    return training_res

training_res = training_log_to_df(training_log)

training_res

#### Best Performance Model 

def get_best_model(training_log_df, list_of_model):
    model_object = None

    best_model_info = training_log_df.sort_values(["r2_score", "training_time"], ascending = [False, True]).iloc[0]
    
    for configuration_data in list_of_model:
        for model_data in list_of_model[configuration_data]:
            if model_data["model_uid"] == best_model_info["model_uid"]:
                model_object = model_data["model_object"]
                break
    
    if model_object == None:
        raise RuntimeError("The best model not found in your list of model.")
    
    return model_object
    

model = get_best_model(training_res, list_of_model)

model

joblib.dump(model, "model/best_vanilla_model.pkl")

y_predd = model.predict(X_test)

y_predd

data_res = {"actual": y_test,
            "pred": y_predd.squeeze()}

mean_squared_error(y_test, y_predd)

mean_absolute_error(y_test, y_predd)

np.sqrt(mean_squared_error(y_test, y_predd))

pd.DataFrame(data_res)
